♻️ Split the input itself to a separate component
Like the text input field, it might be easier if we just use the input field without the wrapped label
